02_finetune_llm_qlora/requirements.txt
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.15.0
trl>=0.7.4
scipy
pyyaml
